\name{eigenclasses-package}
\alias{eigenclasses-package}
\alias{eigenclasses}
\docType{package}
\title{
\packageTitle{eigenclasses}
}
\description{
\packageDescription{eigenclasses}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{eigenclasses}
\packageIndices{eigenclasses}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{eigenclasses}

Maintainer: \packageMaintainer{eigenclasses}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
